HomeWork#4 create Credentials separate entity. Remove credentials from SubjectQuiz
